+++ src/main/java/nextstep/subway/path/dto/PathResponse.java
@@ -1,6 +1,5 @@
 package nextstep.subway.path.dto;
 
-import nextstep.subway.path.domain.SubwayPath;
 import nextstep.subway.station.dto.StationResponse;
 
 import java.util.List;
@@ -8,10 +7,6 @@ import java.util.List;
 public class PathResponse {
     private final List<StationResponse> stations;
     private final int distance;
-
-    public PathResponse(SubwayPath subwayPath) {
-        this(StationResponse.ofList(subwayPath.stations()), subwayPath.distance());
-    }
 
     public PathResponse(List<StationResponse> stations, int distance) {
         this.stations = stations;
